skills/skill_image/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Computer Vision
opencv-python-headless==4.8.1.78
Pillow==10.1.0
imagehash==4.3.1

# Monitoring
psutil==5.9.8

# Utilities
numpy==1.24.3
